Finish split Reducer
